requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.15.1
django-allauth==0.61.1
django-cors-headers
gunicorn
whitenoise
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
python-dotenv
dj_database_url
psycopg2-binary